+++ Statistics/Test/KruskalWallis.hs
@@ -8,19 +8,22 @@
 -- Portability : portable
 --
 module Statistics.Test.KruskalWallis
-  ( kruskalWallisRank
+  ( -- * Kruskal-Wallis test
+    kruskalWallisTest
+    -- ** Building blocks
+  , kruskalWallisRank
   , kruskalWallis
-  , kruskalWallisSignificant
-  , kruskalWallisTest
+  , module Statistics.Test.Types
   ) where
 
 import Data.Ord (comparing)
 import Data.Foldable (foldMap)
 import qualified Data.Vector.Unboxed as U
 import Statistics.Function (sort, sortBy, square)
-import Statistics.Distribution (quantile)
+import Statistics.Distribution (complCumulative)
 import Statistics.Distribution.ChiSquared (chiSquared)
-import Statistics.Test.Types (TestResult(..), significant)
+import Statistics.Types
+import Statistics.Test.Types
 import Statistics.Test.Internal (rank)
 import Statistics.Sample
 import qualified Statistics.Sample.Internal as Sample(sum)
@@ -32,7 +35,7 @@ import qualified Statistics.Sample.Internal as Sample(sum)
 --
 -- The samples and values need not to be ordered but the values in the result
 -- are ordered. Assigned ranks (ties are given their average rank).
-kruskalWallisRank :: [Sample] -> [Sample]
+kruskalWallisRank :: (U.Unbox a, Ord a) => [U.Vector a] -> [U.Vector Double]
 kruskalWallisRank samples = groupByTags
                           . sortBy (comparing fst)
                           . U.zip tags
@@ -54,7 +57,7 @@ kruskalWallisRank samples = groupByTags
 --
 -- In textbooks the output value is usually represented by 'K' or 'H'. This
 -- function already does the ranking.
-kruskalWallis :: [Sample] -> Double
+kruskalWallis :: (U.Unbox a, Ord a) => [U.Vector a] -> Double
 kruskalWallis samples = (nTot - 1) * numerator / denominator
   where
     -- Total number of elements in all samples
@@ -71,29 +74,25 @@ kruskalWallis samples = (nTot - 1) * numerator / denominator
     rsamples = kruskalWallisRank samples
 
 
--- | Calculates whether the Kruskal-Wallis test is significant.
---
--- It uses /Chi-Squared/ distribution for aproximation as long as the sizes are
--- larger than 5. Otherwise the test returns 'Nothing'.
-kruskalWallisSignificant ::
-       [Int]  -- ^ The samples' size
-    -> Double -- ^ The p-value at which to test (e.g. 0.05)
-    -> Double -- ^ K value from 'kruskallWallis'
-    -> Maybe TestResult
-kruskalWallisSignificant ns p k
-    -- Use chi-squared approximation
-    | all (>4) ns = Just . significant $ k > x
-    -- TODO: Implement critical value calculation: kruskalWallisCriticalValue
-    | otherwise = Nothing
-  where
-    x = quantile (chiSquared (length ns - 1)) (1 - p)
-
 -- | Perform Kruskal-Wallis Test for the given samples and required
 -- significance. For additional information check 'kruskalWallis'. This is just
 -- a helper function.
-kruskalWallisTest :: Double -> [Sample] -> Maybe TestResult
-kruskalWallisTest p samples =
-    kruskalWallisSignificant (map U.length samples) p $ kruskalWallis samples
+--
+-- It uses /Chi-Squared/ distribution for aproximation as long as the sizes are
+-- larger than 5. Otherwise the test returns 'Nothing'.
+kruskalWallisTest :: (Ord a, U.Unbox a) => [U.Vector a] -> Maybe (Test ())
+kruskalWallisTest []      = Nothing
+kruskalWallisTest samples
+  -- We use chi-squared approximation here
+  | all (>4) ns = Just $ Test { testSignificance = pValue $ complCumulative d k
+                              , testStatistics   = k
+                              , testDistribution = ()
+                              }
+  | otherwise   = Nothing
+  where
+    k  = kruskalWallis samples
+    ns = map U.length samples
+    d  = chiSquared (length ns - 1)
 
 -- * Helper functions