+++ tests/tests.hs
@@ -4,6 +4,7 @@ import qualified Tests.Function as Function
 import qualified Tests.KDE as KDE
 import qualified Tests.Matrix as Matrix
 import qualified Tests.NonParametric as NonParametric
+import qualified Tests.Parametric as Parametric
 import qualified Tests.Transform as Transform
 import qualified Tests.Correlation as Correlation
 
@@ -13,6 +14,7 @@ main = defaultMain [ Distribution.tests
                    , KDE.tests
                    , Matrix.tests
                    , NonParametric.tests
+                   , Parametric.tests
                    , Transform.tests
                    , Correlation.tests
                    ]